overall code reorganisation
